version=11.5.40_dev draw/axes [x y]
# new feature: draw Vector [x y] or [x y %] {x' y' z'}
# new feature: draw [x y] or [x y %] can be incorporated into lines, curves, arrows
# bug fix: set picking center should not zoom -- just too annoying 
# new feature: set echo myecho [x y] and set echo myecho [x y %]
# new feature: axes position [x y] and axes position [x y %] 
#  sets the axes to the x y screen position or percent indicated.

+++ Jmol/src/org/jmol/shape/BbcageRenderer.java
@@ -24,14 +24,14 @@
 package org.jmol.shape;
 
 import org.jmol.viewer.StateManager;
-import javax.vecmath.Point3i;
+import javax.vecmath.Point3f;
 
 public class BbcageRenderer extends FontLineShapeRenderer {
 
-  final Point3i[] screens = new Point3i[8];
+  final Point3f[] screens = new Point3f[8];
   {
     for (int i = 8; --i >= 0; )
-      screens[i] = new Point3i();
+      screens[i] = new Point3f();
   }
 
   protected void render() {
